requirements.txt
Django==2.0
olefile==0.44
Pillow==4.3.0
pytz==2017.3
virtualenv==15.1.0
